% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent-models.R
\name{AR}
\alias{AR}
\title{Autoregressive Process for Latent Dynamics}
\usage{
AR(order = 1, damp_priors, init_priors, std_prior)
}
\arguments{
\item{order}{Integer. The order of the autoregressive process (p).}

\item{damp_priors}{List of distribution specifications for damping
coefficients. Length must equal \code{order}.}

\item{init_priors}{List of distribution specifications for initial
states. Length must equal \code{order}.}

\item{std_prior}{Distribution specification for innovation standard
deviation.}
}
\value{
An S3 object of class \code{c("epiaware_ar", "epiaware_latent",
  "epiaware_model")} containing:
\describe{
  \item{julia_ref}{Reference to the Julia AR object}
  \item{spec}{List of model specifications}
}
}
\description{
Constructs an autoregressive (AR) latent process of order p for modeling
time-varying log reproduction numbers or other epidemiological parameters.
}
\examples{
\dontrun{
# AR(2) model with truncated normal priors (Mishra et al. 2020)
ar2 <- AR(
  order = 2,
  damp_priors = list(
    truncnorm(0.2, 0.2, 0, 1),
    truncnorm(0.1, 0.05, 0, 1)
  ),
  init_priors = list(norm(0, 0.2), norm(0, 0.2)),
  std_prior = halfnorm(0.1)
)
print(ar2)
}

}
\seealso{
\code{\link{epiaware_call}} for accessing other latent models
}
