% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as_draws_df.epiaware_fit}
\alias{as_draws_df.epiaware_fit}
\title{Convert epiaware_fit to draws_df}
\usage{
\method{as_draws_df}{epiaware_fit}(x, ...)
}
\arguments{
\item{x}{An \code{epiaware_fit} object from \code{fit()}.}

\item{...}{Additional arguments passed to
\code{\link[posterior]{as_draws_df}}.}
}
\value{
A \code{posterior::draws_df} object.
}
\description{
Convert epiaware_fit to draws_df
}
