% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composableIDModelR-package.R
\docType{package}
\name{composableIDModelR-package}
\alias{composableIDModelR}
\alias{composableIDModelR-package}
\title{composableIDModelR: Composable Infectious Disease Models via 'ComposableTuringIDModels.jl'}
\description{
Provides an R interface to 'ComposableTuringIDModels.jl', the composable probabilistic infectious disease modelling package of the 'EpiAware' ecosystem. Models are assembled in R from reusable latent process, infection and observation components, then simulated from and fitted to data in 'Julia' using 'Turing.jl', without requiring 'Julia' expertise.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sbfnk/composableIDModelR}
  \item Report bugs at \url{https://github.com/sbfnk/composableIDModelR/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Funk \email{sebastian.funk@lshtm.ac.uk} [copyright holder]

}
\keyword{internal}
