% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{.check_tspan}
\alias{.check_tspan}
\title{Validate tspan}
\usage{
.check_tspan(tspan, name = "tspan")
}
\arguments{
\item{tspan}{Numeric vector to validate.}

\item{name}{Character string. Name for error messages.}
}
\value{
Invisibly returns \code{TRUE} if valid.
}
\description{
Validate tspan
}
\keyword{internal}
