% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{.cleanup_julia_globals}
\alias{.cleanup_julia_globals}
\title{Clean up Julia global variables}
\usage{
.cleanup_julia_globals()
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Clean up Julia global variables
}
\keyword{internal}
