% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{.discretize_distribution}
\alias{.discretize_distribution}
\title{Discretize a continuous distribution to a PMF}
\usage{
.discretize_distribution(dist_spec, max_days)
}
\description{
Discretize a continuous distribution to a PMF
}
\keyword{internal}
