% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{.ensure_julia_initialized}
\alias{.ensure_julia_initialized}
\title{Initialize Julia lazily (on first use)}
\usage{
.ensure_julia_initialized()
}
\description{
Initialize Julia lazily (on first use)
}
\keyword{internal}
