% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{.find_time_indexed_vars}
\alias{.find_time_indexed_vars}
\title{Find time-indexed variables matching patterns}
\usage{
.find_time_indexed_vars(vars, patterns)
}
\description{
Find time-indexed variables matching patterns
}
\keyword{internal}
