% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.from_julia_array}
\alias{.from_julia_array}
\title{Convert Julia array to R vector}
\usage{
.from_julia_array(julia_obj)
}
\arguments{
\item{julia_obj}{A Julia array object.}
}
\value{
An R vector (passthrough; JuliaCall handles conversion).
}
\description{
Convert Julia array to R vector
}
\keyword{internal}
