% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.julia_chains_to_draws}
\alias{.julia_chains_to_draws}
\title{Convert Turing.jl chains to posterior::draws_df}
\usage{
.julia_chains_to_draws(chains)
}
\description{
Convert Turing.jl chains to posterior::draws_df
}
\keyword{internal}
