% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{.onLoad}
\alias{.onLoad}
\title{Package load hook}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{Character string. Library directory.}

\item{pkgname}{Character string. The package name.}
}
\value{
\code{NULL} (called for side effects).
}
\description{
Package load hook
}
\keyword{internal}
