% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{.plot_latent_trajectory}
\alias{.plot_latent_trajectory}
\title{Plot latent process trajectory}
\usage{
.plot_latent_trajectory(fit, ...)
}
\description{
Plot latent process trajectory
}
\keyword{internal}
