% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{.plot_posterior_distributions}
\alias{.plot_posterior_distributions}
\title{Plot posterior distributions for parameters}
\usage{
.plot_posterior_distributions(fit, ...)
}
\arguments{
\item{fit}{An epiaware_fit object from \code{fit()}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot posterior distributions for parameters
}
\keyword{internal}
