% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{.reconstruct_rt_from_ar}
\alias{.reconstruct_rt_from_ar}
\title{Reconstruct Rt from AR parameters}
\usage{
.reconstruct_rt_from_ar(fit)
}
\description{
Reconstruct Rt from AR parameters
}
\keyword{internal}
