% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia-code.R
\name{.render_string}
\alias{.render_string}
\title{Render a string as an ASCII Julia string literal}
\usage{
.render_string(x)
}
\arguments{
\item{x}{A character scalar.}
}
\value{
A character string.
}
\description{
Render a string as an ASCII Julia string literal
}
\keyword{internal}
