% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{.simulate_renewal}
\alias{.simulate_renewal}
\title{Simulate infections using renewal equation}
\usage{
.simulate_renewal(rt, gen_pmf, init_infections, n_time, observed = NULL)
}
\description{
Simulate infections using renewal equation
}
\keyword{internal}
