% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{.time_axis}
\alias{.time_axis}
\title{Time axis for plotting}
\usage{
.time_axis(fit, n)
}
\arguments{
\item{fit}{An `epiaware_fit` object.}

\item{n}{Number of time points, which may extend beyond the data.}
}
\value{
Dates if the fit has them, otherwise integers.
}
\description{
Time axis for plotting
}
\keyword{internal}
