% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.to_julia_vector}
\alias{.to_julia_vector}
\title{Convert R vector to Julia vector}
\usage{
.to_julia_vector(x)
}
\arguments{
\item{x}{An R vector.}
}
\value{
A Julia-compatible vector (passthrough; JuliaCall handles conversion).
}
\description{
Convert R vector to Julia vector
}
\keyword{internal}
