% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{epiaware_available}
\alias{epiaware_available}
\title{Check whether the Julia backend is running}
\usage{
epiaware_available()
}
\value{
Logical. `TRUE` if Julia has been set up in this R session.
}
\description{
Check whether the Julia backend is running
}
\examples{
epiaware_available()
}
