% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{exponential}
\alias{exponential}
\title{Exponential distribution}
\usage{
exponential(rate)
}
\arguments{
\item{rate}{Numeric. Rate parameter (inverse of mean).}
}
\value{
A list specification for an exponential distribution.
}
\description{
Specifies an exponential distribution.
}
\examples{
\dontrun{
# Exponential prior with mean 10
prior <- exponential(1 / 10)
}

}
