% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{halfnorm}
\alias{halfnorm}
\title{Half-normal distribution}
\usage{
halfnorm(sd)
}
\arguments{
\item{sd}{Numeric. Scale parameter (standard deviation of underlying normal).}
}
\value{
A list specification for a half-normal distribution.
}
\description{
Specifies a half-normal distribution (normal distribution truncated to
positive values).
}
\examples{
\dontrun{
# Prior for standard deviation parameters
sigma_prior <- halfnorm(0.1)
}

}
