% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{lognorm}
\alias{lognorm}
\title{Log-normal distribution}
\usage{
lognorm(meanlog, sdlog)
}
\arguments{
\item{meanlog}{Numeric. Mean of the distribution on the log scale.}

\item{sdlog}{Numeric. Standard deviation of the distribution on the log
scale.}
}
\value{
A list specification for a log-normal distribution.
}
\description{
Specifies a log-normal distribution.
}
\examples{
\dontrun{
# Delay distribution
delay <- lognorm(1.6, 0.42)
}

}
