% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation-models.R
\name{NegativeBinomialError}
\alias{NegativeBinomialError}
\title{Negative Binomial Observation Error Model}
\usage{
NegativeBinomialError(cluster_factor_prior)
}
\arguments{
\item{cluster_factor_prior}{Distribution specification for the cluster factor
(sqrt(1/phi)), which represents the coefficient of variation of observation
noise.}
}
\value{
An S3 object of class \code{c("epiaware_negbin",
  "epiaware_observation", "epiaware_model")} containing:
\describe{
  \item{julia_ref}{Reference to the Julia NegativeBinomialError object}
  \item{spec}{List of model specifications}
}
}
\description{
Links latent infections to observed case counts using a negative binomial
distribution. The cluster_factor parameterizes the overdispersion as the
coefficient of variation (sqrt(1/phi)), which is more intuitive for setting
priors than the dispersion parameter phi directly.
}
\examples{
\dontrun{
# Negative binomial observation model
negbin <- NegativeBinomialError(
  cluster_factor_prior = halfnorm(0.1)
)
print(negbin)
}

}
