% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{norm}
\alias{norm}
\title{Normal distribution}
\usage{
norm(mean, sd)
}
\arguments{
\item{mean}{Numeric. Mean of the distribution.}

\item{sd}{Numeric. Standard deviation of the distribution.}
}
\value{
A list specification for a normal distribution.
}
\description{
Specifies a normal (Gaussian) distribution.
}
\examples{
\dontrun{
# Standard normal
prior1 <- norm(0, 1)

# Prior for log Rt
prior2 <- norm(log(1.5), 0.2)
}

}
