% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent-models.R
\name{print.epiaware_ar}
\alias{print.epiaware_ar}
\title{Print method for AR latent models}
\usage{
\method{print}{epiaware_ar}(x, ...)
}
\arguments{
\item{x}{An \code{epiaware_ar} object.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Print method for AR latent models
}
