% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation-models.R
\name{print.epiaware_delay}
\alias{print.epiaware_delay}
\title{Print method for latent delay observation models}
\usage{
\method{print}{epiaware_delay}(x, ...)
}
\arguments{
\item{x}{An \code{epiaware_delay} object.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Print method for latent delay observation models
}
