% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.epiaware_fit}
\alias{print.epiaware_fit}
\title{Print Method for Fitted EpiAware Models}
\usage{
\method{print}{epiaware_fit}(x, ...)
}
\arguments{
\item{x}{An \code{epiaware_fit} object from \code{fit()}.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Print Method for Fitted EpiAware Models
}
