% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.epiaware_generic}
\alias{print.epiaware_generic}
\title{Print method for generic EpiAware calls}
\usage{
\method{print}{epiaware_generic}(x, ...)
}
\arguments{
\item{x}{An \code{epiaware_generic} object.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Print method for generic EpiAware calls
}
