% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation-models.R
\name{print.epiaware_negbin}
\alias{print.epiaware_negbin}
\title{Print method for negative binomial observation models}
\usage{
\method{print}{epiaware_negbin}(x, ...)
}
\arguments{
\item{x}{An \code{epiaware_negbin} object.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Print method for negative binomial observation models
}
