% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{print.epiaware_nuts}
\alias{print.epiaware_nuts}
\title{Print method for NUTS sampler configuration}
\usage{
\method{print}{epiaware_nuts}(x, ...)
}
\arguments{
\item{x}{An \code{epiaware_nuts} object.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Print method for NUTS sampler configuration
}
