% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi-problem.R
\name{print.epiaware_problem}
\alias{print.epiaware_problem}
\title{Print method for EpiAware problems}
\usage{
\method{print}{epiaware_problem}(x, ...)
}
\arguments{
\item{x}{An \code{epiaware_problem} object.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Print method for EpiAware problems
}
