% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection-models.R
\name{print.epiaware_renewal}
\alias{print.epiaware_renewal}
\title{Print method for renewal infection models}
\usage{
\method{print}{epiaware_renewal}(x, ...)
}
\arguments{
\item{x}{An \code{epiaware_renewal} object.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Print method for renewal infection models
}
