% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{Normal}
\alias{Gamma}
\alias{LogNormal}
\alias{Exponential}
\alias{Weibull}
\alias{Beta}
\alias{Fixed}
\alias{NonParametric}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{distspec}{\code{\link[distspec]{Beta}}, \code{\link[distspec]{Exponential}}, \code{\link[distspec]{Fixed}}, \code{\link[distspec]{Gamma}}, \code{\link[distspec]{LogNormal}}, \code{\link[distspec]{NonParametric}}, \code{\link[distspec]{Normal}}, \code{\link[distspec]{Weibull}}}
}}

