% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{truncnorm}
\alias{truncnorm}
\title{Truncated normal distribution}
\usage{
truncnorm(mean, sd, lower, upper)
}
\arguments{
\item{mean}{Numeric. Mean of the underlying normal distribution.}

\item{sd}{Numeric. Standard deviation of the underlying normal distribution.}

\item{lower}{Numeric. Lower truncation bound.}

\item{upper}{Numeric. Upper truncation bound.}
}
\value{
A list specification for a truncated normal distribution.
}
\description{
Specifies a normal distribution truncated to a specified range.
}
\examples{
\dontrun{
# AR damping coefficient bounded to [0, 1]
damp_prior <- truncnorm(0.8, 0.2, 0, 1)
}

}
