% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{.as_draws}
\alias{.as_draws}
\title{Convert bridge output to a draws_df}
\usage{
.as_draws(values, names, index)
}
\arguments{
\item{values}{Matrix of draws (rows) by variables (columns).}

\item{names}{Character vector of variable names.}

\item{index}{List with `.chain` and `.iteration` vectors.}
}
\value{
A [posterior::draws_df].
}
\description{
Convert bridge output to a draws_df
}
\keyword{internal}
