% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-convert.R
\name{.as_prior}
\alias{.as_prior}
\title{Convert a distribution used as a prior to a Julia component}
\usage{
.as_prior(x, arg_name = deparse(substitute(x)))
}
\arguments{
\item{x}{A distspec distribution or an composableIDModelR component.}

\item{arg_name}{Name used in error messages.}
}
\value{
A component of class `cidm_distribution`.
}
\description{
Priors are sampled by the model, so their parameters must be fixed.
}
\keyword{internal}
