% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-convert.R
\name{.assert_single}
\alias{.assert_single}
\title{Check that a distribution is a single distribution}
\usage{
.assert_single(x, arg_name)
}
\arguments{
\item{x}{A distspec distribution.}

\item{arg_name}{Name used in error messages.}
}
\value{
Invisibly `TRUE`.
}
\description{
Check that a distribution is a single distribution
}
\keyword{internal}
