% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{.bridge}
\alias{.bridge}
\title{Call a bridge function, starting Julia if needed}
\usage{
.bridge(fn, ...)
}
\arguments{
\item{fn}{Name of a function in the `EpiAwareR` Julia module.}

\item{...}{Arguments passed to the function.}
}
\value{
The translated result of the call.
}
\description{
Releases the Julia objects of fits that R has garbage collected before
making the call. Releasing waits until here because finalisers may run
while another Julia call is in progress.
}
\keyword{internal}
