% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{.check_distribution}
\alias{.check_distribution}
\title{Validate distribution specification}
\usage{
.check_distribution(x, name = deparse(substitute(x)), null_ok = FALSE)
}
\description{
Validate distribution specification
}
\keyword{internal}
