% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{.check_distribution_list}
\alias{.check_distribution_list}
\title{Validate list of distribution specifications}
\usage{
.check_distribution_list(x, name = deparse(substitute(x)), len = NULL)
}
\description{
Validate list of distribution specifications
}
\keyword{internal}
