% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-convert.R
\name{.dist_max}
\alias{.dist_max}
\title{Maximum of a distspec distribution}
\usage{
.dist_max(x)
}
\arguments{
\item{x}{A distspec distribution.}
}
\value{
The maximum, or \code{Inf} if it is unbounded.
}
\description{
Maximum of a distspec distribution
}
\keyword{internal}
