% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{.generate_quantities}
\alias{.generate_quantities}
\title{Generate quantities from fitted model}
\usage{
.generate_quantities(julia_model, draws_obj)
}
\description{
Generate quantities from fitted model
}
\keyword{internal}
