% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{.get_latent_plot_settings}
\alias{.get_latent_plot_settings}
\title{Get plot settings based on infection model type}
\usage{
.get_latent_plot_settings(fit)
}
\description{
Get plot settings based on infection model type
}
\keyword{internal}
