% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.is_distribution_list}
\alias{.is_distribution_list}
\title{Check if a list contains only distribution specifications}
\usage{
.is_distribution_list(x)
}
\description{
Check if a list contains only distribution specifications
}
\keyword{internal}
