% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{.plot_latent_trajectory}
\alias{.plot_latent_trajectory}
\title{Plot latent process trajectory}
\usage{
.plot_latent_trajectory(fit, ...)
}
\arguments{
\item{fit}{An \code{epiaware_fit} object from \code{fit()}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot latent process trajectory
}
\keyword{internal}
