% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{.plot_posterior_predictive}
\alias{.plot_posterior_predictive}
\title{Plot posterior predictive for cases}
\usage{
.plot_posterior_predictive(fit, ...)
}
\description{
Plot posterior predictive for cases
}
\keyword{internal}
