% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.prepare_data_for_julia}
\alias{.prepare_data_for_julia}
\title{Prepare R data frame for Julia}
\usage{
.prepare_data_for_julia(data, tspan)
}
\description{
Prepare R data frame for Julia
}
\keyword{internal}
