% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.to_julia_dist}
\alias{.to_julia_dist}
\title{Convert R distribution specification to Julia Distribution object}
\usage{
.to_julia_dist(dist_spec)
}
\description{
Convert R distribution specification to Julia Distribution object
}
\keyword{internal}
