% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.to_julia_dist_vector}
\alias{.to_julia_dist_vector}
\title{Convert R list of Julia distributions to Julia vector}
\usage{
.to_julia_dist_vector(dist_list)
}
\description{
Convert R list of Julia distributions to Julia vector
}
\keyword{internal}
