% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{epiaware_available}
\alias{epiaware_available}
\title{Check if Julia and EpiAware are available}
\usage{
epiaware_available()
}
\value{
Logical. TRUE if Julia and EpiAware are available, FALSE otherwise.
}
\description{
Tests whether Julia is configured and EpiAware packages are accessible.
}
\examples{
\dontrun{
if (epiaware_available()) {
  # Run EpiAware analysis
} else {
  epiaware_setup_julia()
}
}

}
