% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{epiaware_setup_julia}
\alias{epiaware_setup_julia}
\title{Setup Julia and EpiAware}
\usage{
epiaware_setup_julia(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical. If TRUE, prints progress messages.}
}
\value{
Invisible TRUE on success.
}
\description{
Configures Julia and installs required Julia packages for EpiAwareR.
If a `julia` version is pinned in upstream EpiAware's Project.toml, that
version is installed via [juliaup](https://github.com/JuliaLang/juliaup)
and used; otherwise the system Julia is used.
}
\details{
Heavy lifting (subprocess install, lazy-init guard) is delegated to
[juliaready::julia_ready()].
}
\examples{
\dontrun{
epiaware_setup_julia()
}

}
