% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{epiaware_setup_julia}
\alias{epiaware_setup_julia}
\title{Setup Julia and EpiAware}
\usage{
epiaware_setup_julia(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical. If TRUE, prints progress messages. Default is TRUE.}
}
\value{
Invisible TRUE if setup succeeds, otherwise throws an error.
}
\description{
Configures Julia and installs required Julia packages for EpiAwareR.
This function should be run on first use if automatic setup fails.
}
\examples{
\dontrun{
# Setup Julia and install EpiAware packages
epiaware_setup_julia()
}

}
