% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiAwareR-package.R
\docType{package}
\name{EpiAwareR-package}
\alias{EpiAwareR}
\alias{EpiAwareR-package}
\title{EpiAwareR: R Interface to the 'EpiAware' Infectious Disease Modelling Framework}
\description{
Provides an R interface to the 'Julia'-based 'EpiAware' compositional infectious disease modelling framework. Enables users to build flexible epidemiological models by composing reusable components for latent processes, infection dynamics, and observation models without requiring 'Julia' expertise.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sbfnk/EpiAwareR}
  \item Report bugs at \url{https://github.com/sbfnk/EpiAwareR/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Funk \email{sebastian.funk@lshtm.ac.uk} [copyright holder]

}
\keyword{internal}
